// src/types/api-types/utils/Itslearning.RestApi.Entities.InstantMessageRecipient.ts
export type ItslearningRestApiEntitiesInstantMessageRecipient = {
  TypeName: string;
  Id: number;
  TypeNameId: string;
  Label: string;
  EncodedLabel: string;
  Organization: string;
  ShortLabel: string;
  SearchLabel: string;
  SortField: number;
  ParticipantsCount: number;
  ProfileImageUrl: string;
  ColorSchema: string;
  Code: string;
  CanReceive: boolean;
  RoleInHierarchyTitle: string;
};